+++ test/integration/util.ts
@@ -2,16 +2,22 @@ require("dotenv").config();
 process.env = {
     ethereum: "http://127.0.0.1:8545",
     node_http: "http://127.0.0.1:8000/daostack/graphql",
+    node_ws: "http://graph-node:8001/by-name/daostack",
     test_mnemonic:
         "behave pipe turkey animal voyage dial relief menu blush match jeans general",
     ...process.env
 };
 
+const { execute } = require('apollo-link');
+const { WebSocketLink } = require('apollo-link-ws');
+const { SubscriptionClient } = require('subscriptions-transport-ws');
+const ws = require('ws');
+
 import axios from "axios";
 const Web3 = require("web3");
 import * as HDWallet from "hdwallet-accounts";
 
-const { node_http, ethereum, test_mnemonic } = process.env;
+const { node_ws,node_http, ethereum, test_mnemonic } = process.env;
 
 export async function query(q: string, maxDelay = 1000) {
     await new Promise((res, rej) => setTimeout(res, maxDelay));
@@ -60,3 +66,16 @@ export function padZeros(str: string, max = 36) {
     str = str.toString();
     return str.length < max ? padZeros("0" + str, max) : str;
 }
+
+
+const getWsClient = function(wsurl) {
+  const client = new SubscriptionClient(
+    wsurl, {reconnect: true}, ws
+  );
+  return client;
+};
+
+export const createSubscriptionObservable = (query, variables = 0 ,wsurl = node_ws) => {
+  const link = new WebSocketLink(getWsClient(wsurl));
+  return execute(link, {query: query, variables: variables});
+};
